[Clang] Allow malloc and alloc_size attributes for functions returning structs
These attributes restricted the possible function signatures to the ones
returning a pointer, which is not the case for some custom / proposed
allocation functions. For example, P0901R11 proposed ::operator new
overloads that return a return_size_t result - a struct that contains
a pointer to the allocated memory as well as the actual size of the
allocated memory. A similar lib function is __size_returning_new.

Relax the mentioned restriction and allow "malloc" and "alloc_size"
attributes to be applied not just on the functions returning pointer
types, but also on the functions returning records whose first member is
a pointer (which would be assumed to point to the allocated memory).
This is the case for return_size_t as well as std::span, should it be
returned from an annotated function.

As the previous restriction was dictated by the need to set the noalias
attribute, adjust the code generation logic to only assign it to the
return values of the functions that return a pointer type.

In a future change, codegen can be improved to recognize the
noalias-ness of the pointer within the returned structs, which is not
done in this patch.

This change also helps unlock the alloc token instrumentation for such
non-standard allocation functions:
https://clang.llvm.org/docs/AllocToken.html#instrumenting-non-standard-allocation-functions