[Clang] Allow malloc and alloc_size attributes for functions returning structs
These attributes restrict the possible function signatures to the ones
returning a pointer, which is not the case for some non-standard
allocation functions variants. For example, P0901R11 proposed ::operator new
overloads that return a return_size_t result - a struct that contains
a pointer to the allocated memory as well as the actual size of the
allocated memory. Another example is __size_returning_new.

Relax the mentioned restriction and allow "malloc" and "alloc_size"
attributes to be applied not just to the functions returning pointer
types, but also to the functions returning records whose first member is
a pointer (which would be assumed to point to the allocated memory).
This is the case for return_size_t as well as std::span, should it be
returned from such an annotated function.

As the previous restriction was dictated by the need to set the noalias
attribute, adjust the code generation logic to only assign the attribute
when the returned value type is a pointer type.

In future commits, codegen can be improved to recognize the
noalias-ness of the pointer returned inside a return_size_t-like struct.

This change also helps unlock the alloc token instrumentation for such
non-standard allocation functions:
https://clang.llvm.org/docs/AllocToken.html#instrumenting-non-standard-allocation-functions