[Clang] Allow malloc and alloc_size attributes for functions returning structs
The allocation token functionality relies on the instrumentation of the
allocation functions, which added a new use case for the "malloc" and
"alloc_size" attributes.

These attributes restricted the possible function signatures to the ones
returning a pointer, which is not the case for some custom / proposed
allocation functions. For example, P0901R11 proposed ::operator new
overloads that return a return_size_t result - a struct that contains
a pointer to the allocated memory as well as the actual size of the
allocated memory.

Relax the mentioned restriction and allow "malloc" and "alloc_size"
attributes to be applied not just on the functions returning pointer
types, but also on the functions returning records whose first member is
a pointer (which would be assumed to point to the allocated memory).

As the previous restriction was dictated by the need to set the noalias
attribute, adjust the code generation logic to only assign it to the
return values of the functions that return a pointer type.