[AllocToken] Test alloc token instrumentation for size-returning functions
The allocation token functionality relies on the instrumentation of the
allocation functions, which added a new use case for the "malloc" and
"size_alloc" attributes.

With the new "malloc_span" attribute for annotating functions returning
structs that wrap the pointer to the allocated data, this commit adjusts
the allocation token test to ensure that __size_returning_new variants
are correctly handled by the alloc token instrumentation.