[AllocToken] Test alloc token instrumentation for size-returning functions
The allocation token functionality relies on the instrumentation of the
allocation functions, which added a new use case for the "malloc" and
"size_alloc" attributes.

With the attributes now supporting functions that return a structs that
wrap the pointer, this commit adjusts the allocation token test to ensure
that __size_returning_new variants are now also included into the alloc
token instrumentation.