[Clang] Propagate noalias from malloc attribute when returned via struct
This commit follows up on the "[Clang] Allow malloc and alloc_size attributes
for functions returning structs" which relaxed the
__attribute__((malloc)) to apply to functions returning std::span-like
structs.

This patch implements the deferred functionality of actually ensuring
the `noalias` property of the pointers returned by the malloc-like
functions.

To achieve this, a new function attribute is introduced:
`returns_noalias_field="N"`: incidates that the field at index N of the
returned struct is a `noalias` pointer.

The attribute is set on the `__attribute__((malloc))` functions whose
return value is a std::span-like struct.

The commit changes the alias analysis logic to recognize and apply the
returns_noalias_field attribute in the simple scenario of doing
extractvalue on a struct returned by a function that is marked by this
attribute.