[Clang] Propagate noalias from malloc_span attribute
This commit follows up on the "XXX" which introduced the new
__attribute__((malloc_span)) to apply to functions returning span-like
structs.

This patch implements the deferred functionality of actually ensuring
the `noalias` property of the pointers returned by the malloc-like
functions.

To achieve this, a new function attribute is introduced:
`returns_noalias_field="N"`: incidates that the field at index N of the
returned struct is a `noalias` pointer.

The commit changes the alias analysis logic to recognize and apply the
returns_noalias_field attribute in the simple scenario of doing
extractvalue on a struct returned by a function that is marked by this
attribute.